requirements.txt
autoflake==1.4
black==21.12b0
click==8.0.3
flake8==4.0.1
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.22.1
pandas==1.3.5
pathspec==0.9.0
Pillow==9.0.0
pip-autoremove==0.10.0
platformdirs==2.4.1
protobuf==3.19.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
tensorboardX==2.4.1
tomli==1.2.3
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.10.1+cpu
torchaudio==0.10.1+cpu
torchvision==0.11.2+cpu
tqdm==4.62.3
typing_extensions==4.0.1
wandb==0.12.9
 pytorch_lightning==1.5.9